feat: Bump nix channel version and fix tmux nord theme by pinning (for fixing go-task completion)
`go-task` completion is failing because of an excessively stale version,
seeing as this is the correct way to do thing going forward we just bump
the nix channel to get this to work.

Upon bumping I discover that the tmux nord theme didn't work so fix that
by pinning. I am unsure why it didn't work but I cba to find out.